requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
google-generativeai==0.5.2
python-dotenv==1.0.1
rich==13.7.1
